Working CNN
Model is already trained, If you want to train it again you need to set load_model in line 58 to false and set the number of epochs you want in line 57.\nThe model ouputs an image which has to be inside the repository, preferably in the folder /test_imgs/ and the name of the file included in line 118 in main. After the image with the label of either a cat or a dog is closed you will get the output of the evaluation of the model in the form of a JSON. Inside this JSON is the model name, model loss and model accuracy respectively.